refactor: 移动 Menu 脚本
